requirements.txt
pyyaml
pygame
pytest
psutil